pyproject.toml
[project]
name = "nhl-led-scoreboard"
version = "2025.1.0"
description = "🚨 Display NHL score of your favorite team on a Raspberry Pi driven RGB LED matrix. 🚨"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "backoff>=2.2.1",
    "cairosvg>=2.6.0",
    "dbus-python>=1.3.2",
    "diskcache>=5.6.3",
    "env-canada>=0.6.1",
    "ephem>=4.1.6",
    "fastjsonschema>=2.14.4",
    "geopy>=2.4.1",
    "gpiozero==1.6.2",
    "ipinfo>=5.1.1",
    "iso6709>=0.1.5",
    "lastversion>=1.1.6",
    "nameparser>=1.0.6",
    "nhl-api-py>=2.18.0",
    "noaa-sdk>=0.1.18",
    "paho-mqtt>=2.1.0",
    "pillow>=10.0.0",
    "python-tsl2591==0.2.0",
    "regex>=2020.4.4",
    "rgbmatrix",
    "rgbmatrixemulator>=0.12.1",
    "rich>=13.9.4",
    "tk>=0.1.0",
    "tzlocal>=5.2",
    "holidays",
]

[tool.uv.sources]
rgbmatrix = { path = "submodules/matrix/bindings/python" }
